promote ui-slot-composer.ts to just be slot-composer.ts
