+++ shells/smoke-shell/smoke.html
@@ -11,7 +11,7 @@
 <script type="module">
   import {App} from './app.js';
   import {Runtime} from '../../build/runtime/runtime.js';
-  import {UiSlotComposer} from '../../build/runtime/ui-slot-composer.js';
+  import {SlotComposer} from '../../build/runtime/slot-composer.js';
   import {attachRenderer} from '../lib/xen-renderer.js';
 
   // run
@@ -20,7 +20,7 @@
       // configure arcs environment
       Runtime.init('../..');
       // construct renderer
-      const composer = new UiSlotComposer();
+      const composer = new SlotComposer();
       composer.root = document.body;
       attachRenderer(composer).dispatch = (pid, eventlet) => {
         console.log('ui-broker/composer dispatch for pid', pid, eventlet);
